urban_point_cloud_analyzer/scripts/train.py
#!/usr/bin/env python3
"""
Training script for Urban Point Cloud Analyzer
"""
import argparse
import os
import sys
import yaml
from pathlib import Path

import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader

# Add the project root to the path
sys.path.insert(0, str(Path(__file__).resolve().parent.parent))

from urban_point_cloud_analyzer.data.loaders import DALESDataset
from urban_point_cloud_analyzer.models import get_model
from urban_point_cloud_analyzer.training import get_loss_function, get_optimizer, get_scheduler
from urban_point_cloud_analyzer.utils.logger import setup_logger


def parse_args():
    parser = argparse.ArgumentParser(description="Train Urban Point Cloud Analyzer")
    parser.add_argument("--config", type=str, default="configs/default_config.yaml", help="Path to config file")
    parser.add_argument("--data_dir", type=str, help="Path to data directory")
    parser.add_argument("--output_dir", type=str, default="outputs", help="Output directory")
    parser.add_argument("--resume", type=str, help="Path to checkpoint to resume from")
    parser.add_argument("--seed", type=int, default=42, help="Random seed")
    return parser.parse_args()


def main():
    args = parse_args()
    
    # Load config
    with open(args.config, 'r') as f:
        config = yaml.safe_load(f)
    
    # Override config with command line arguments
    if args.data_dir:
        config['data']['data_path'] = args.data_dir
    
    # Setup output directory
    output_dir = Path(args.output_dir)
    output_dir.mkdir(exist_ok=True, parents=True)
    
    # Setup logger
    logger = setup_logger(output_dir / "train.log")
    logger.info(f"Config: {config}")
    
    # Set random seed
    torch.manual_seed(args.seed)
    
    # Setup device
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    logger.info(f"Using device: {device}")
    
    # Create dataset and dataloader
    logger.info("Creating datasets...")
    # TODO: Implement actual dataset and dataloader
    
    # Create model
    logger.info("Creating model...")
    # TODO: Implement model creation
    
    # Define loss function, optimizer and scheduler
    logger.info("Setting up training...")
    # TODO: Implement training setup
    
    # Training loop
    logger.info("Starting training...")
    # TODO: Implement training loop
    
    logger.info("Training completed!")


if __name__ == "__main__":
    main()
